Init: MP-to-Policy Happy/Unhappy path functional tests
