+++ tox.ini
@@ -20,8 +20,8 @@ install_command = python -m pip install {opts} {packages}
 setenv = VIRTUAL_ENV={envdir}
          LANGUAGE=en_US
          LC_ALL=en_US.utf-8
-         OS_STDOUT_CAPTURE=0
-         OS_STDERR_CAPTURE=0
+         OS_STDOUT_CAPTURE=1
+         OS_STDERR_CAPTURE=1
          OS_TEST_TIMEOUT=720
          DEBUG={env:NSXV3_LOGIN_HOSTNAME:False}
          NSXV3_LOGIN_HOSTNAME={env:NSXV3_LOGIN_HOSTNAME:}
@@ -68,13 +68,11 @@ commands =
 
 [testenv:functional]
 basepython = python3.6
-commands =
-    env TEST_OSPROFILER=1 pytest networking_nsxv3/tests/functional/
+commands = env TEST_OSPROFILER=1 pytest networking_nsxv3/tests/functional/
 
 [testenv:mp2policy]
 basepython = python3.6
-commands =
-    env TEST_OSPROFILER=1 pytest -s networking_nsxv3/tests/functional/test_mp2policy_brownfield.py
+commands = env TEST_OSPROFILER=1 pytest networking_nsxv3/tests/functional/mp2policy/
 
 [testenv:pep8]
 basepython = python3.6
